Added Lucene storage provider

+++ tileserver/src/main/java/com/saltaku/tileserver/providers/basemaps/BasemapCompressor.java
@@ -1,6 +1,6 @@
 package com.saltaku.tileserver.providers.basemaps;
 
 public interface BasemapCompressor {
-public int[] compress(int[] in);
-public int[] decompress(int[] in);
+public byte[] compress(int[] in);
+public int[] decompress(byte[] in);
 }
